PDB Dataset

+++ ProteinPairsGenerator/datasets/pdb_dataset.py
@@ -14,11 +14,11 @@ def base(data_pdb: AtomGroup) -> Data:
 
     # Find intersequence distance
     n = len(seq)
-    ids = torch.arange(0, n).unsqueeze(0).unsqueeze(0)
+    ids = torch.arange(0, n, dtype=torch.float32).unsqueeze(0).unsqueeze(0)
     seq_distances = torch.cdist(ids, ids).squeeze(0).squeeze(0)
 
     # Compute caresian distances
-    coords = torch.from_numpy(data_pdb.getCoords())
+    coords = torch.from_numpy(data_pdb.getCoords(), dtype=torch.float32)
     cart_distances = torch.cdist(coords.unsqueeze(0), coords.unsqueeze(0)).squeeze(0)
 
     # Mask and put in correct shape
